# gen/go/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: goods-center-logic服务文档
  title: goods-center-logic API
  version: "1.0"
servers:
- url: http://127.0.0.1:3013/goodsCenterLogic
paths:
  /v1/common/generateId:
    get:
      description: 生成id-描述
      parameters:
      - description: 认证信息 eg:xxxx-xxxx-xxxx-xxx
        in: header
        name: Authorization
        schema:
          type: string
      - description: 生成id数量 1-1000
        in: query
        name: num
        required: true
        schema:
          type: integer
      - description: ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/internal_api_http_servicev1.HttpGenerateIDResponse'
          description: 请求成功
      summary: 雪花ID生成
      tags:
      - 公共接口
  /v1/exampleGet:
    get:
      description: get接口示例
      parameters:
      - description: create_time
        in: query
        name: create_time
        required: true
        schema:
          type: integer
      - description: query_id
        in: query
        name: query_id
        required: true
        schema:
          type: string
      - description: user_id 用户id
        in: query
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/internal_api_http_dto.ExampleGetResponse'
          description: 请求成功
      summary: get接口示例
      tags:
      - Example
  /v1/exampleGetOne:
    get:
      description: getOne接口示例
      parameters:
      - description: create_time
        in: query
        name: create_time
        required: true
        schema:
          type: integer
      - description: query_id
        in: query
        name: query_id
        required: true
        schema:
          type: string
      - description: user_id 用户id
        in: query
        name: user_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/internal_api_http_dto.ExampleGetOneResponse'
          description: 请求成功
      summary: getOne接口示例
      tags:
      - Example
  /v1/examplePost:
    post:
      description: "\nios购买项类型 <a href=\"https://developer.apple.com/documentation/appstoreconnectapi/list_all_in-app_purchases_for_an_app\"\
        > 详情</a> </br> \n\n\nandroid订阅 <a href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/monetization.subscriptions/list?hl=zh-cn\"\
        > 详情</a> </br> \n\nandroid非订阅<a href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/inappproducts/list?hl=zh-cn\"\
        > 详情</a> </br> \n\nandroid订阅产品的类型<a href=\"https://developers.google.com/android-publisher/api-ref/rest/v3/inappproducts?hl=zh-cn#PurchaseType\"\
        > 详情</a> </br>"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_api_http_dto.ExamplePostRequest'
        description: 请求参数
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_http_dto.ExamplePostResponse'
          description: 请求成功
          headers:
            Token:
              description: token
              schema:
                type: string
            Token2:
              description: token2
              schema:
                type: string
            Location:
              description: /entity/1
              schema:
                type: string
      summary: post 接口 示例
      tags:
      - Example
      x-codegen-request-body-name: body
components:
  schemas:
    internal_api_http_dto.ExampleGetOneResponse:
      example:
        msg: success
        traceId: traceId
        code: 1
        data:
          country: country
          vip_info: 0
          user_id: user_id
          last_login: last_login
      properties:
        code:
          description: "code:  0 成功; 非0失败;"
          example: 1
          type: integer
        data:
          $ref: '#/components/schemas/internal_api_http_dto.UserPortraitData'
        msg:
          description: 错误消息
          example: success
          type: string
        traceId:
          description: traceId
          type: string
      required:
      - code
      - data
      - msg
      - traceId
      type: object
    internal_api_http_dto.ExampleGetResponse:
      example:
        msg: success
        traceId: traceId
        code: 1
        data:
          country: country
          vip_info: 0
          user_id: user_id
          last_login: last_login
      properties:
        code:
          description: "code:  0 成功; 非0失败;"
          example: 1
          type: integer
        data:
          $ref: '#/components/schemas/internal_api_http_dto.UserPortraitData'
        msg:
          description: 错误消息
          example: success
          type: string
        traceId:
          description: traceId
          type: string
      required:
      - code
      - data
      - msg
      - traceId
      type: object
    internal_api_http_dto.ExamplePostRequest:
      properties:
        app_id:
          description: AppID 应用id
          type: string
        c_ver:
          description: CVer 客户端协议版本号
          type: string
        cmd_id:
          description: |-
            CmdId 功能类型枚举:
            1 AUTH鉴权;
            2 QA 问答；
            4 总结SUMMARY;
            5 KEYWORDS 关键词；
            6 CLASSIFICATION 文档分类；
            7 RENAME  重命名；
            8 mindmap 思维导图；
          example: 1
          type: integer
        fid:
          description: Fid： 文档id
          type: string
        ignore_feature_prompt:
          description: IgnoreFeaturePrompt 是否忽略功能检查  1 是； 0 否；
          type: integer
        lang:
          description: Lang 语言
          type: string
        question:
          description: Question 问题
          type: string
        show_gpt_2:
          description: show_gpt_2
          type: string
        token:
          description: Token 鉴权token
          example: sdfsdsdfsd
          type: string
      required:
      - app_id
      - c_ver
      - cmd_id
      - fid
      - lang
      - question
      - token
      type: object
    internal_api_http_dto.ExamplePostResponse:
      example:
        msg: success
        traceId: traceId
        code: 1
        data:
          country: country
          vip_info: 0
          user_id: user_id
          last_login: last_login
      properties:
        code:
          description: "code:  0 成功; 非0失败;"
          example: 1
          type: integer
        data:
          $ref: '#/components/schemas/internal_api_http_dto.UserPortraitData'
        msg:
          description: 错误消息
          example: success
          type: string
        traceId:
          description: traceId
          type: string
      required:
      - code
      - data
      - msg
      - traceId
      type: object
    internal_api_http_dto.UserPortraitData:
      example:
        country: country
        vip_info: 0
        user_id: user_id
        last_login: last_login
      properties:
        country:
          description: 国家
          type: string
        last_login:
          description: 上次登陆时间
          type: string
        user_id:
          description: 用户id
          type: string
        vip_info:
          description: 是否为VIP，0/1
          type: integer
      type: object
    internal_api_http_servicev1.HttpGenerateIDResponse:
      example:
        msg: success
        traceId: traceId
        code: 1
        data:
        - data
        - data
      properties:
        code:
          description: "code:  0 成功; 非0失败;"
          example: 1
          type: integer
        data:
          description: 生成id数组
          items:
            type: string
          type: array
        msg:
          description: 错误消息
          example: success
          type: string
        traceId:
          description: traceId
          type: string
      required:
      - code
      - data
      - msg
      - traceId
      type: object
x-original-swagger-version: "2.0"